pyproject.toml
[project]
name = "prompt-attribution"
version = "0.1.0"
description = "Debugging prompt attribution via ablation testing"
authors = [
    {name = "Greyson Lampley"}
]
readme = "README.md"
requires-python = "^3.9"
dependencies = [
    "openai (>=1.78.1,<2.0.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "rich (>=14.0.0,<15.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
ruff = "^0.11.9"
pytest = "^8.3.5"
mypy = "^1.15.0"
pre-commit = "^4.2.0"

